refactor: move profiles to canonical package
