+++ tests/test_python_package_skeleton.py
@@ -67,8 +67,11 @@ PACKAGE_MODULES = (
     "easyuse_anima.nodes.regional_nodes",
     "easyuse_anima.nodes.sam3_nodes",
     "easyuse_anima.profiles",
+    "easyuse_anima.profiles.aio",
     "easyuse_anima.profiles.contract",
+    "easyuse_anima.profiles.lora",
     "easyuse_anima.profiles.mutation",
+    "easyuse_anima.profiles.repository",
     "easyuse_anima.settings",
     "easyuse_anima.settings.repository",
     "easyuse_anima.settings.schema",
