+++ tests/test_registry_scanner_safety.py
@@ -68,8 +68,11 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/nodes/sam3_nodes.py",
     "easyuse_anima/nodes/wildcard_nodes.py",
     "easyuse_anima/profiles/__init__.py",
+    "easyuse_anima/profiles/aio.py",
     "easyuse_anima/profiles/contract.py",
+    "easyuse_anima/profiles/lora.py",
     "easyuse_anima/profiles/mutation.py",
+    "easyuse_anima/profiles/repository.py",
     "easyuse_anima/settings/__init__.py",
     "easyuse_anima/settings/repository.py",
     "easyuse_anima/settings/schema.py",
